# openapi.yaml
openapi: 3.0.1
info:
  title: finance data plugin
  description: A plugin that allows you to retrieve financial data for a company based on its ticker symbol.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /financial_ratios/{ticker}:
    get:
      operationId: getFinancialRatios
      summary: Retrieve financial ratios for a company based on its ticker symbol
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    # Add properties here based on the structure of the JSON response
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string

  /stock_news/{ticker}:
    get:
      operationId: getStockNews
      summary: Retrieve recent stock news for a company based on its ticker symbol
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    # Add properties here based on the structure of the JSON response
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string


  /earning_call_transcript/{ticker}:
    get:
      operationId: getEarningCallTranscript
      summary: Retrieve recent quater earning call transcript
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
              
                    # Add properties here based on the structure of the JSON response
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string

  /insider_trading/{ticker}:
    get:
      operationId: getInsiderTrading
      summary: Retrieve recent trading activities of insiders in the company (e.g. whether employees like senior engieers at a company sell/buy their company's stock)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    # Add properties here based on the structure of the JSON response
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string

  /esg_score/{ticker}:
    get:
      operationId: getESGScore
      summary: Retrieve recent ESG score of a company (i.e score on how much a company cares about environment, social, and governance)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    # Add properties here based on the structure of the JSON response
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string

  /SEC_Filling/{ticker}:
    get:
      operationId: getSECFilling
      summary: Retrieve recent SEC filling of a company, it only returns the link to the filling
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    # Add properties here based on the structure of the JSON response
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string

  /revenue_product_segmentation_annually/{ticker}:
    get:
      operationId: getRevenueProductSegmentationAnnually
      summary: Retrieve company revenue segmentation by product annually
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /sales_revenue_by_product_segement_quarterly/{ticker}:
    get:
      operationId: getRevenueProductSegmentationQuarterly
      summary: Retrieve company revenue segmentation by product quarterly
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /sales_revenue_by_geographic_segmentation_annually/{ticker}:
    get:
      operationId: getRevenueGeographicSegmentationAnnually
      summary: Retrieve company revenue segmentation by geographic annually
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /sales_revenue_by_geographic_segmentation_quarterly/{ticker}:
    get:
      operationId: getRevenueGeographicSegmentationQuarterly
      summary: Retrieve company revenue segmentation by geographic quarterly
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /social_media_sentiments/{ticker}:
    get:
      operationId: getSocialMediaSentiments
      summary: Retrieve social media sentiments of a company (including twitter, stocktwits, etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /company_profile/{ticker}:
    get:
      operationId: getCompanyProfile
      summary: Retrieve company profile (e.g. company description, industry, price, beta, ceo, employee number, address, contact, logo etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /company_executives/{ticker}:
    get:
      operationId: getCompanyExecutives
      summary: Retrieve company executives (e.g. CEO, CFO, CTO, etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /company_quote/{ticker}:
    get:
      operationId: getCompanyQuote
      summary: Retrieve company quote (e.g. price, market cap, enterprise value, volume, avg volume, etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /company_dcf/{ticker}:
    get:
      operationId: getCompanyDCFResult
      summary: Retrieve company DCF result (e.g. this one only gvies result ro retrieve the DCF, cash flow and wacc details use the endpoint detailed_DCF_projection_including_wacc_data)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:



  /detailed_dcf_projection_including_wacc_data/{ticker}:
    get:
      operationId: getCompanyDCFDetails
      summary: Retrieve company DCF details (e.g. DCF projection, free cash flow, wacc, etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /income_statement_quarter/{ticker}:
    get:
      operationId: getIncomeStatementQuarterly
      summary: Retrieve company Income Statement quarterly (e.g. revenue, gross profit, net income, etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /balance_sheet_quarter/{ticker}:
    get:
      operationId: getBalanceSheetQuarterly
      summary: Retrieve company Balance Sheet quarterly (e.g. cash, debt, assets, liabilities, etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:

  /cash_flow_quarter/{ticker}:
    get:
      operationId: getCashFlowQuarterly
      summary: Retrieve company cash flow quarterly (e.g. operating cash flow, investing cash flow, financing cash flow, etc.)
      parameters:
        - in: path
          name: ticker
          schema:
            type: string
          required: true
          description: The ticker symbol of the company for which information are to be fetched.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:


components:
  schemas:
    getTodosResponse:
      type: object
      properties:
        todos:
          type: array
          items:
            type: string
          description: The list of todos.
    addTodoRequest:
      type: object
      required:
        - todo
      properties:
        todo:
          type: string
          description: The todo to add to the list.
    deleteTodoRequest:
      type: object
      required:
        - todo_idx
      properties:
        todo_idx:
          type: integer
          description: The index of the todo to delete.